Move `py` variable from Daemon.kt to App.kt:
On startup, if ACRA attempts to resend a crash report which previously failed, it accesses `py` to get the Python version number. But this also initializes all the other global variables in the same file, which caused the following exception:

java.lang.IllegalStateException: Cannot invoke setValue on a background thread
        at androidx.lifecycle.LiveData.assertMainThread(LiveData.java:443)
        at androidx.lifecycle.LiveData.setValue(LiveData.java:286)
        at androidx.lifecycle.MutableLiveData.setValue(MutableLiveData.java:33)
        at org.electroncash.electroncash3.DaemonKt.<clinit>(Daemon.kt:23)
        at org.electroncash.electroncash3.DaemonKt.getPy(Unknown Source:0) 
        at org.electroncash.electroncash3.CrashKt$reportToJson$4.invoke(Crash.kt:102)
        at org.electroncash.electroncash3.CrashKt$reportToJson$4.invoke(Unknown Source:0)
        at org.electroncash.electroncash3.CrashKt.putJson(Crash.kt:116)
        at org.electroncash.electroncash3.CrashKt.reportToJson(Crash.kt:101)
        at org.electroncash.electroncash3.CrashhubSender.convertToString(Crash.kt:62)
        at org.acra.sender.HttpSender.send(HttpSender.java:127)
        at org.acra.sender.ReportDistributor.sendCrashReport(ReportDistributor.java:111)
        at org.acra.sender.ReportDistributor.distribute(ReportDistributor.java:76)
        at org.acra.sender.SenderService.onHandleWork(SenderService.java:84)
        at androidx.core.app.JobIntentService$CommandProcessor.doInBackground(JobIntentService.java:392)
        at androidx.core.app.JobIntentService$CommandProcessor.doInBackground(JobIntentService.java:383)
        at android.os.AsyncTask$2.call(AsyncTask.java:333)
        at java.util.concurrent.FutureTask.run(FutureTask.java:266)
        at java.util.concurrent.ThreadPoolExecutor.runWorker(ThreadPoolExecutor.java:1167) 
        at java.util.concurrent.ThreadPoolExecutor$Worker.run(ThreadPoolExecutor.java:641) 
        at java.lang.Thread.run(Thread.java:764)
