+++ android/app/src/main/java/org/electroncash/electroncash3/Daemon.kt
@@ -4,17 +4,8 @@ import androidx.lifecycle.MutableLiveData
 import android.widget.Toast
 import com.chaquo.python.PyException
 import com.chaquo.python.PyObject
-import com.chaquo.python.Python
-import com.chaquo.python.android.AndroidPlatform
 
 
-val py by lazy {
-    Python.start(AndroidPlatform(app))
-    Python.getInstance()
-}
-fun libMod(name: String) = py.getModule("electroncash.$name")!!
-fun guiMod(name: String) = py.getModule("electroncash_gui.android.$name")!!
-val libNetworks by lazy { libMod("networks") }
 val guiDaemon by lazy { guiMod("daemon") }
 
 val WATCHDOG_INTERVAL = 1000L
